src/main/java/com/hu3diger/order_processor/services/AuthenticationService.java
package com.hu3diger.order_processor.services;
import com.hu3diger.order_processor.dtos.LoginResponseDto;
import com.hu3diger.order_processor.dtos.LoginUserDto;
import com.hu3diger.order_processor.dtos.RegisterUserDto;
import com.hu3diger.order_processor.entities.UserEntity;
import com.hu3diger.order_processor.repositories.UserRepository;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class AuthenticationService {

    private final JwtService jwtService;

    private final UserRepository userRepository;

    private final PasswordEncoder passwordEncoder;

    private final AuthenticationManager authenticationManager;

    public AuthenticationService(
            UserRepository userRepository,
            AuthenticationManager authenticationManager,
            PasswordEncoder passwordEncoder,
            JwtService jwtService
    ) {
        this.authenticationManager = authenticationManager;
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
        this.jwtService = jwtService;
    }

    public UserEntity signup(RegisterUserDto input) {
        var user = new UserEntity();
        user.setFullName(input.getFullName());
        user.setEmail(input.getEmail());
        user.setPassword(passwordEncoder.encode(input.getPassword()));

        return userRepository.save(user);
    }


    public UserEntity authenticate(LoginUserDto input) {
        authenticationManager.authenticate(
                new UsernamePasswordAuthenticationToken(
                        input.getEmail(),
                        input.getPassword()
                )
        );

        return userRepository.findByEmail(input.getEmail())
                .orElseThrow();
    }

    public LoginResponseDto authenticateAndGetResponse(LoginUserDto input) {
        var authenticatedUser = this.authenticate(input);

        String jwtToken = jwtService.generateToken(authenticatedUser);

        var loginResponse = new LoginResponseDto();
        loginResponse.setToken(jwtToken);
        loginResponse.setExpiresIn(jwtService.getExpirationTime());

        return loginResponse;
    }
}
